test(contract): add a test for contract deployment
